NOICE

+++ Reaper.Engine/WorldObjectBounds.cs
@@ -15,8 +15,8 @@ namespace Reaper.Engine
 
         public float X;
         public float Y;
-        public int Width;
-        public int Height;
+        public float Width;
+        public float Height;
 
         public float Left => X;
         public float Right => X + Width;
@@ -27,7 +27,7 @@ namespace Reaper.Engine
 
         public Rectangle ToRectangle()
         {
-            return new Rectangle((int)X, (int)Y, Width, Height);
+            return new Rectangle((int)X, (int)Y, (int)Width, (int)Height);
         }
 
         public void Offset(float x, float y)
